# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stash-box metadata database",
        "title": "Stash-box REST API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Stash App",
            "url": "https://github.com/stashapp/stash-box"
        },
        "license": {
            "name": "AGPL-3.0",
            "url": "https://www.gnu.org/licenses/agpl-3.0.en.html"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/studios": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "List studios with filters",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort field (name, created_at, updated_at)",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort direction (asc, desc)",
                        "name": "direction",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Studio"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "Create a new studio",
                "parameters": [
                    {
                        "description": "Studio data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.StudioCreateInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Studio"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/studios/by-name/{name}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "Get studio by name",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Studio name",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Studio"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/studios/search": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "Search studios by term",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "term",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 25,
                        "description": "Result limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Studio"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/studios/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "Get studio by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Studio ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Studio"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "Update a studio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Studio ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated studio data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.StudioUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_models.Studio"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "Delete a studio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Studio ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/studios/{id}/favorite": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "Favorite a studio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Studio ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth SessionAuth": []
                    }
                ],
                "tags": [
                    "studios"
                ],
                "summary": "Unfavorite a studio",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Studio ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "additionalProperties": {
                                                "type": "boolean"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_stashapp_stash-box_internal_api_rest_handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_stashapp_stash-box_internal_api_rest_handlers.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total_count": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_stashapp_stash-box_internal_api_rest_handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {}
            }
        },
...
        "github_com_stashapp_stash-box_internal_models.Studio": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted": {
                    "type": "boolean"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "parent_studio_id": {
                    "$ref": "#/definitions/github_com_gofrs_uuid.NullUUID"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication",
            "type": "apiKey",
            "name": "ApiKey",
            "in": "header"
        },
        "SessionAuth": {
            "description": "Session cookie authentication",
            "type": "apiKey",
            "name": "stashbox",
            "in": "cookie"
        }
    }
}